#	Assets/Animators/Magnet knight run left.aseprite.meta
#	Assets/Animators/flying enemy side movement.png.meta
#	Assets/Animators/magnet knight run back.aseprite.meta
#	Assets/Animators/magnet knight run front.aseprite.meta
#	Assets/Animators/magnet knight run right.aseprite.meta
#	Assets/Animators/scrap 2 design finale.aseprite.meta

+++ Assets/David/Script/StatsHUD.cs
@@ -14,8 +14,7 @@ public class StatsHUD : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        healthBar.fillAmount = stats.Health / stats.MaxHealth;
-        ScrapBar.fillAmount = stats.Scrap / stats.MaxScrap;
-        
+        healthBar.fillAmount = (float)stats.Health / stats.MaxHealth;
+        ScrapBar.fillAmount = (float)stats.Scrap / stats.MaxScrap;
     }
 }
